+++ src/styles/components/Timer.styled.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const StyledTimer = styled.div`
+  display: inline-block;
+  outline: 1px solid #EEE;
+`;
+
+export default StyledTimer;